src/main/java/com/github/cc3002/finalreality/model/character/player/Engineer.java
package com.github.cc3002.finalreality.model.character.player;

import com.github.cc3002.finalreality.model.character.ICharacter;
import com.github.cc3002.finalreality.model.weapon.AxeWeapon;
import com.github.cc3002.finalreality.model.weapon.BowWeapon;
import com.github.cc3002.finalreality.model.weapon.Weapon;
import com.github.cc3002.finalreality.model.weapon.WeaponType;
import org.jetbrains.annotations.NotNull;

import java.util.concurrent.BlockingQueue;

public class Engineer extends AbstractPlayerCharacter{
    /**
     * Creates a new character that it's class is engineer.
     *
     * @param name
     *     the character's name
     * @param turnsQueue
     *     the queue with the characters waiting for their turn
     */
    public Engineer(@NotNull String name, @NotNull BlockingQueue<ICharacter> turnsQueue) {
        super(name, turnsQueue, CharacterClass.ENGINEER);
    }

    /**
     * Equips an Axe to the Engineer
     * @param axename
     *      Axe's name
     */
    public void equipAxe(String axename){
        this.equippedWeapon = new AxeWeapon(axename);
    }
    /**
     * Equips an Bow to the Engineer
     * @param bowname
     *      Bow's name
     */
    public void equipBow(String bowname) {
        this.equippedWeapon = new BowWeapon(bowname);
    }
}
